Add retries for recoverable Octokit errors
Currently the entire job will fail if a single request to the GitHub API
results in an error. By using the [retry
plugin](https://github.com/octokit/plugin-retry.js) it will by default
retry for 3 attempts whenever a "recoverable" error is encounted. These
are configurable and the defaults can be seen here:

https://github.com/octokit/plugin-retry.js/blob/dcf1fa38337b06d1a82cda596e8839306498f8c3/src/index.ts#L12-L15